Date trunc implementation: clippy fix

+++ rust/datafusion/src/physical_plan/datetime_expressions.rs
@@ -233,7 +233,7 @@ pub fn date_trunc(args: &[ArrayRef]) -> Result<TimestampNanosecondArray> {
     let result = range
         .map(|i| {
             if array.is_null(i) {
-                Ok(0 as i64)
+                Ok(0_i64)
             } else {
                 let date_time = match granularity_array.value(i) {
                     "second" => array
